fact_check_system/requirements.txt
langgraph>=0.0.20
langchain>=0.0.309
langchain-openai>=0.0.2
langchain-anthropic>=0.0.9
langchain-google-genai>=0.0.5
pydantic>=2.5.0
requests>=2.31.0
python-dotenv>=1.0.0
tavily-python>=0.2.4
spacy>=3.7.0
nltk>=3.8
beautifulsoup4>=4.12.0
newsapi-python>=0.2.7
fastapi>=0.100.0
uvicorn>=0.24.0
google-generativeai>=0.3.0 